Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/form-submission-0/form-data-set-empty-file.window-expected.txt


FAIL Empty <input type=file> is still serialized assert_equals: type expected "application/octet-stream" but got ""
